feat(amlume-shop): add Uber nullaway + Google Error Prone
NullAway is a tool to help eliminate NullPointerExceptions (NPEs) in your Java code. To use NullAway, first add @Nullable annotations in your code wherever a field, method parameter, or return value may be null. Given these annotations, NullAway performs a series of type-based, local checks to ensure that any pointer that gets dereferenced in your code cannot be null. NullAway is similar to the type-based nullability checking in the Kotlin and Swift languages, and the Checker Framework and Eradicate null checkers for Java.

Error Prone …
- hooks into your standard build, so all developers run it without thinking
- tells you about mistakes immediately after they’re made
- produces suggested fixes, allowing you to build tooling on it