refactor: use json and improve timeout for UX
As we are using redis to store session data, we have to use json for serialization. We should not store passkey data in cache, we should store it in database. The passkey challenge is stored in cache.

The timeout for the passkey registration must be adequate in a way the user have a nice experience. Considering the fact that the user has to register a passkey for the first time, we might need to use the timeout of 1 or 2 minutes.