refactor: [ReturnValueIgnored] Return value of 'valueOf' must be used
From Error Prone: That warning, [ReturnValueIgnored] Return value of 'valueOf' must be used, is flagging that the result of Enum.valueOf(enumClass, value.toUpperCase()); isn't being assigned to anything.

In the isValidEnum method, we're correctly using a try-catch block to determine if Enum.valueOf succeeds (meaning the string is a valid enum constant) or fails (throws IllegalArgumentException). The return value of Enum.valueOf itself isn't strictly needed for the logic of isValidEnum because we're only interested in whether an exception is thrown.

However, Error Prone prefers that if a method returns a value, that value should be acknowledged in some way, even if it's just assigned to an unused variable. This can sometimes help catch subtle bugs where a developer might have intended to use the return value but forgot.